Consume from egress topic and forward to available
Rather than listening for a success vaildation, we can instead simply
listen to the host-egresss topic. If a message appears here, it means
that inventory was able to do everything it needs and pass it along to
an announcement. At this point we're just republishing anything from
egress on to avaialble, but in the way the messages are expected over
there.

This effectively turns the validation queue into a rejected queue since
all we use that for is to know if we need to reject a file to another
bucket.