[refactor] Refactor Kafka SSL
Need to refactor SSL so that we can decouple it from sasl. Moved the
security protocol out of the sasl config since fedramp will have it
without authentication.

Refactored the way we get the kafka config to reduce the amount of
repeated code that we're using in the consumer and producer.

Signed-off-by: Stephen Adams <tsadams@gmail.com>